fix(ios): spell the cancel-slider opacity so Xcode 16.4 can type it
`.opacity(1 + dragX / 110)` is ambiguous to the Swift 5 type checker in
Xcode 16.4. Newer toolchains resolve it, which is why it has compiled on
every developer machine and only broke the first time anything built the
app on the CI toolchain.

The line is unchanged since before this PR — it is a pre-existing
incompatibility with the version CI pins, not a regression here. It went
unnoticed because until #518 merged, nothing in CI compiled
`ios/bitchat.xcodeproj` at all.

Found by the iOS build job on its first run against a real PR, which is
exactly the job it was added to do.

+++ ios/bitchat/Views/Sonar/SonarComponents.swift
@@ -4255,7 +4255,12 @@ struct SNComposer: View {
                         .font(SonarTheme.uiFont(size: 12))
                         .foregroundColor(cancelArmed ? SonarTheme.danger : SonarTheme.text3)
                 }
-                .opacity(1 + dragX / 110)
+                // Spelled with explicit `CGFloat`s: the literal form
+                // (`1 + dragX / 110`) is ambiguous to Swift 5 in Xcode 16.4,
+                // which is what CI builds with, while newer toolchains resolve
+                // it. It compiled locally for everyone and broke the first time
+                // anything built the app on the CI toolchain.
+                .opacity(CGFloat(1) + dragX / CGFloat(110))
             }
             .padding(.vertical, 7)
             .padding(.horizontal, 12)
